# [2.x] Bring media assets into the Hyde kernel