--- a/packages/framework/src/Foundation/Concerns/HasMediaFiles.php
+++ b/packages/framework/src/Foundation/Concerns/HasMediaFiles.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Hyde\Foundation\Concerns;
+
+/**
+ * @internal Single-use trait for the Filesystem class.
+ *
+ * @see \Hyde\Foundation\Kernel\Filesystem
+ */
+trait HasMediaFiles
+{
+    //
+}