--- a/packages/framework/src/Foundation/Concerns/HasMediaFiles.php
+++ b/packages/framework/src/Foundation/Concerns/HasMediaFiles.php
@@ -13,4 +13,14 @@ trait HasMediaFiles
 {
     /** @return array<string, \Hyde\Support\Filesystem\MediaFile> The array keys are the filenames relative to the _media/ directory */
     protected array $assets = [];
+
+    /**
+     * Get all media files in the project.
+     *
+     * @return array<string, \Hyde\Support\Filesystem\MediaFile>
+     */
+    public function assets(): array
+    {
+        return $this->assets;
+    }
 }