--- a/packages/framework/src/Foundation/Concerns/HasMediaFiles.php
+++ b/packages/framework/src/Foundation/Concerns/HasMediaFiles.php
@@ -11,5 +11,6 @@ namespace Hyde\Foundation\Concerns;
  */
 trait HasMediaFiles
 {
-    //
+    /** @return array<string, \Hyde\Support\Filesystem\MediaFile> The array keys are the filenames relative to the _media/ directory */
+    protected array $assets = [];
 }