--- a/packages/framework/src/Foundation/Concerns/HasMediaFiles.php
+++ b/packages/framework/src/Foundation/Concerns/HasMediaFiles.php
@@ -12,7 +12,7 @@ namespace Hyde\Foundation\Concerns;
 trait HasMediaFiles
 {
     /** @return array<string, \Hyde\Support\Filesystem\MediaFile> The array keys are the filenames relative to the _media/ directory */
-    protected array $assets = [];
+    protected array $assets;
 
     /**
      * Get all media files in the project.