--- a/packages/framework/tests/Feature/DiscoveryServiceTest.php
+++ b/packages/framework/tests/Feature/DiscoveryServiceTest.php
@@ -88,10 +88,12 @@ class DiscoveryServiceTest extends UnitTestCase
     {
         $testFiles = ['png', 'svg', 'jpg', 'jpeg', 'gif', 'ico', 'css', 'js'];
 
-        foreach ($testFiles as $fileType) {
-            $path = 'test.'.$fileType;
-            $this->file('_media/'.$path);
-            $this->assertContains($path, MediaFile::files());
+        foreach ($testFiles as $type) {
+            $this->file("_media/test.$type");
+        }
+
+        foreach ($testFiles as $type) {
+            $this->assertContains("test.$type", MediaFile::files());
         }
     }