--- a/packages/framework/tests/Feature/Foundation/FilesystemTest.php
+++ b/packages/framework/tests/Feature/Foundation/FilesystemTest.php
@@ -21,6 +21,7 @@ use function Hyde\normalize_slashes;
 /**
  * @covers \Hyde\Foundation\HydeKernel
  * @covers \Hyde\Foundation\Kernel\Filesystem
+ * @covers \Hyde\Foundation\Concerns\HasMediaFiles
  */
 class FilesystemTest extends UnitTestCase
 {