--- a/packages/framework/tests/Feature/HydeKernelTest.php
+++ b/packages/framework/tests/Feature/HydeKernelTest.php
@@ -204,6 +204,34 @@ class HydeKernelTest extends TestCase
         $this->assertSame('assets/foo.jpg', Hyde::asset('foo.jpg'));
     }
 
+    public function testAssetsHelperGetsAllSiteAssets()
+    {
+        $this->assertEquals(new Collection([
+            'app.css' => new MediaFile('_media/app.css'),
+        ]), Hyde::assets());
+    }
+
+    public function testAssetsHelperGetsAllSiteAssetsAsArray()
+    {
+        $assets = Hyde::assets()->toArray();
+
+        $assets['app.css']['length'] = 123;
+
+        $this->assertSame([
+            'app.css' => [
+                'name' => 'app.css',
+                'path' => '_media/app.css',
+                'length' => 123,
+                'mimeType' => 'text/css',
+            ],
+        ], $assets);
+    }
+
+    public function testAssetsHelperReturnsAssetCollectionSingleton()
+    {
+        $this->assertSame(Hyde::assets(), Hyde::assets());
+    }
+
     public function testRouteHelper()
     {
         $this->assertNotNull(Hyde::route('index'));