This is part of https://github.com/hydephp/develop/pull/1904

>  See if we can cache the media assets during discovery. A good location may be in the kernel Filesystem instance. This would also provide a small performance optimization if we create cache busting keys at discovery time. It would also improve DX with almost no performance penalty as we could throw an exception when trying to get a non-existent file.

Quick benchmarking shows this now goes from `~2.5ms` per `MediaFile::files()` call to `~2.7ms` for the first call, and essentially zero at all subsequent calls in the request lifecycle. This is at the cost of the kernel needing to rerun discovery if media files change during the request, which should only happen in unit testing that are coupled to the filesystem.