@caendesilva 2024-07-28
Okay, added a more robust benchmarker in , with about 15 files, including nested ones, and a 10MB one.

```yaml
# Before:
Warmup: 15.8036ms
Benchmark: 10.2105987ms # avg/1000/its

# After:
Warmup: 16.9699ms
Benchmark: 0.00093500000000001ms # avg/1000/its
```